Added git merge error handling in script gitRelease

+++ gitRelease.sh
@@ -2,11 +2,16 @@
 #Pre-requisites: all commits on develop done
 if [ $# != 1 ]
 	then
-		echo "No version ID provided"
+		echo "No version ID provided."
 		exit -1
 fi
 git checkout release
 git merge develop --no-commit
+if [ $? == 1]
+	then
+		echo "Merge conflicts to be solved."
+		exit -1
+fi
 rm Version.h~
 mv Version.h Version.h~
 sed s/"#define VERSION \"[^\"]*\""/"#define VERSION \"$1\""/ <Version.h~ >Version.h
